Add stateful hastache template function to retrieve CIDR blocks.
Initial support is added to use mustache template function:

{{#next_cidr_block}}{{/next_cidr_block}}

On each use of this function, the value will be replaced with the
next CIDR block available. Currently this value comes from a
generated list of address ranges, but it will be changed to check
in an AWS account to provide a CIDR block that is not yet in use.